refactor(lazy-pages): join memory pages realisations
